# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "nanocloud",
    "description": "A CLI tool for managing your Nanocloud",
    "license": {
      "name": ""
    },
    "version": "0.0.1"
  },
  "paths": {
...
    "/apis/nanocloud.io/v1/namespaces/{namespace}/bundles/{name}/actions/restart": {
      "post": {
        "tags": [
          "nanocloud"
        ],
        "operationId": "restart_bundle",
        "parameters": [
          {
            "name": "namespace",
            "in": "path",
            "description": "Bundle namespace",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "description": "Bundle name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Restart accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceActionResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "NanocloudCertificate": []
          },
          {
            "NanocloudBearer": [
              "services.manage"
            ]
          }
        ]
      }
    },
    "/apis/nanocloud.io/v1/namespaces/{namespace}/bundles/{name}/actions/start": {
      "post": {
        "tags": [
          "nanocloud"
        ],
        "operationId": "start_bundle",
        "parameters": [
          {
            "name": "namespace",
            "in": "path",
            "description": "Bundle namespace",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "description": "Bundle name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Start accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceActionResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "NanocloudCertificate": []
          },
          {
            "NanocloudBearer": [
              "services.manage"
            ]
          }
        ]
      }
    },
    "/apis/nanocloud.io/v1/namespaces/{namespace}/bundles/{name}/actions/stop": {
      "post": {
        "tags": [
          "nanocloud"
        ],
        "operationId": "stop_bundle",
        "parameters": [
          {
            "name": "namespace",
            "in": "path",
            "description": "Bundle namespace",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "description": "Bundle name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Stop accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceActionResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "NanocloudCertificate": []
          },
          {
            "NanocloudBearer": [
              "services.manage"
            ]
          }
        ]
      }
    },
    "/apis/nanocloud.io/v1/namespaces/{namespace}/bundles/{name}/actions/uninstall": {
      "post": {
        "tags": [
          "nanocloud"
        ],
        "operationId": "uninstall_bundle",
        "parameters": [
          {
            "name": "namespace",
            "in": "path",
            "description": "Bundle namespace",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "description": "Bundle name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Nanocloud-Owner",
            "in": "header",
            "description": "Backup owner identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Uninstall accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceActionResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "NanocloudCertificate": []
          },
          {
            "NanocloudBearer": [
              "services.manage"
            ]
          }
        ]
      }
    },
    "/apis/nanocloud.io/v1/namespaces/{namespace}/bundles/{name}/backups/latest": {
      "get": {
        "tags": [
          "nanocloud"
        ],
        "operationId": "stream_latest_backup",
        "parameters": [
          {
            "name": "namespace",
            "in": "path",
            "description": "Bundle namespace",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "description": "Bundle name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Nanocloud-Owner",
            "in": "header",
            "description": "Backup owner identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Latest backup archive"
          },
          "404": {
            "description": "No backup available",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "NanocloudCertificate": []
          },
          {
            "NanocloudBearer": [
              "services.read"
            ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApplyConflict": {
        "type": "object",
        "required": [
          "path",
          "existingManager"
        ],
        "properties": {
          "existingManager": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "conflicts": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/ApplyConflict"
            }
          },
          "error": {
            "type": "string"
          }
        }
      },
...
      "ServiceActionResponse": {
        "type": "object",
        "required": [
          "service",
          "action"
        ],
        "properties": {
          "action": {
            "type": "string"
          },
          "namespace": {
            "type": [
              "string",
              "null"
            ]
          },
          "service": {
            "type": "string"
          },
          "started": {
            "type": [
              "boolean",
              "null"
            ]
          }
        }
      },
...
    }
  },
...
}